pyproject.toml
[project]
name = "udemy-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "flask>=3.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.10",
    "langchainhub>=0.1.21",
    "langsmith>=0.4.4",
    "python-dotenv>=1.1.1",
    "tweepy>=4.16.0",
]
